Remove configuration page and migrate to MetaBrainz.MusicBrainz.CoverArt
