--- Jellyfin.Plugin.CoverArtArchive/Constants.cs
@@ -1,18 +1,0 @@
-namespace Jellyfin.Plugin.CoverArtArchive
-{
-    /// <summary>
-    /// Constants.
-    /// </summary>
-    public static class Constants
-    {
-        /// <summary>
-        /// Gets the user agent.
-        /// </summary>
-        public const string UserAgent = "jellyfin-plugin-coverartarchive";
-
-        /// <summary>
-        /// Gets the api base url.
-        /// </summary>
-        public const string ApiBaseUrl = "https://coverartarchive.org";
-    }
-}
