+++ build.yaml
@@ -12,6 +12,8 @@ category: "Metadata"
 owner: "jellyfin"
 artifacts:
   - "Jellyfin.Plugin.CoverArtArchive.dll"
+  - "MetaBrainz.Common.Json.dll"
+  - "MetaBrainz.MusicBrainz.CoverArt.dll"
 changelog: |2-
   ### New features and improvements ###
   - chore: plugin images (#5) @h1dden-da3m0n
